[STDPRO-62] :construction:

+++ bin/alpha.js
@@ -1,10 +1,9 @@
 #! /usr/bin/env node
 const { program } = require("commander");
-const { mkdirSync, readFileSync, writeFileSync, readdirSync, readdir } = require("fs");
+const { mkdirSync, readFileSync, writeFileSync, readdirSync, copyFile } = require("fs");
 const process = require("process");
 const child_process = require("child_process");
 const path = require("path");
-const neededDirs = ["src", "public", "src\\components"];
 const extendPkg = {
 	scripts: {
 		dev: "vite",
@@ -12,11 +11,13 @@ const extendPkg = {
 		serve: "vite preview",
 	},
 	dependencies: {
+		"ant-design-vue": "^3.2.15",
+		less: "^4.1.3",
+		"less-loader": "^11.1.0",
+		pinia: "^2.0.33",
+		uuid: "^9.0.0",
 		vue: "^3.2.47",
 		"vue-router": "^4.1.6",
-		vuex: "^4.1.0",
-		"ant-design-vue": "^3.2.15",
-		uuid: "^9.0.0",
 	},
 	devDependencies: {
 		"@rushstack/eslint-patch": "^1.2.0",
@@ -28,15 +29,18 @@ const extendPkg = {
 		"@vue/eslint-config-typescript": "^11.0.2",
 		"@vue/test-utils": "^2.3.0",
 		"@vue/tsconfig": "^0.1.3",
-		"unplugin-vue-components": "^0.22.9",
-		"vite-plugin-commonjs": "^0.6.1",
 		eslint: "^8.34.0",
 		"eslint-plugin-vue": "^9.9.0",
 		jsdom: "^21.1.0",
 		"npm-run-all": "^4.1.5",
 		prettier: "^2.8.4",
 		typescript: "~4.8.4",
-		vite: "^4.1.4",
+		"unplugin-vue-components": "^0.22.9",
+		vite: "^3.2.5",
+		"vite-plugin-commonjs": "^0.6.1",
+		"vite-plugin-compression": "^0.5.1",
+		"vite-plugin-optimize-persist": "^0.1.2",
+		"vite-plugin-package-config": "^0.1.1",
 		vitest: "^0.29.1",
 		"vue-tsc": "^1.2.0",
 	},
@@ -94,29 +98,34 @@ function extendPackage(extend) {
 	});
 }
 
-function createDir(targetPath) {
-	neededDirs.forEach((dir) => {
-		mkdirSync(`${targetPath}\\${dir}`, { recursive: true });
-	});
+function createStructure(targetPath) {
+	const resolvedStructurePath = path.join(__dirname, "../projectStructure");
+	copyFileTo(targetPath, resolvedStructurePath);
 }
 
 function copyFileTo(targetPath, srcPath) {
-	readdirSync(srcPath).forEach((filename) => {
-		const ds = path.join(srcPath, filename);
-		const contentBuffer = readFileSync(ds);
-		const file = `${targetPath}\\${filename}`;
-		writeFileSync(file, contentBuffer, {
-			flag: "w+",
-		});
-		console.log(`File:${file} created.`);
+	const result = readdirSync(srcPath, { withFileTypes: true });
+	result.forEach((fileOrDir) => {
+		if (fileOrDir.isFile()) {
+			const ds = path.join(srcPath, fileOrDir.name);
+			const contentBuffer = readFileSync(ds);
+			const file = `${targetPath}\\${fileOrDir.name}`;
+			writeFileSync(file, contentBuffer, {
+				flag: "w+",
+			});
+			console.log(`File:${file} created.`);
+		} else {
+			const _targetPath = `${targetPath}\\${fileOrDir.name}`;
+			const _srcPath = `${srcPath}\\${fileOrDir.name}`;
+			mkdirSync(_targetPath);
+			copyFileTo(_targetPath, _srcPath);
+		}
 	});
 }
 
-function createFile(targetPath) {
-	const resolvedConfigPath = path.join(__dirname, "../configTmp");
-	const resolvedComponentPath = path.join(__dirname, "../components");
+function createConfig(targetPath) {
+	const resolvedConfigPath = path.join(__dirname, "../projectConfig");
 	copyFileTo(targetPath, resolvedConfigPath);
-	copyFileTo(`${targetPath}\\src\\components`, resolvedComponentPath);
 	writeFileSync(
 		`${targetPath}\\.gitignore`,
 		`node_modules
@@ -135,9 +144,19 @@ function init(initPath) {
 	try {
 		executeCommand("npm init -y", initPath);
 		extendPackage(extendPkg);
-		executeCommand(`npm install`, initPath);
-		createDir(initPath);
-		createFile(initPath);
+		// executeCommand(`npm install --legacy-peer-deps`, initPath);
+		createStructure(initPath);
+		createConfig(initPath);
+		const layoutModuleUrl = "http://git.dqalpha.com/frontend/components/web/layout_vue_web.git";
+		const layoutPath = "./src/layout";
+		const routerModuleUrl = "http://git.dqalpha.com/frontend/components/web/router_vue_web.git";
+		const routerPath = "./src/router";
+		const alphaComponentModuleUrl = "http://git.dqalpha.com/frontend/components/web/alpha-components-antd.git";
+		const alphaComponentPath = "./src/components/alpha-components-antd";
+		executeCommand(`git -C ${initPath} init`);
+		executeCommand(`git -C ${initPath} submodule add ${layoutModuleUrl} ${layoutPath}`);
+		executeCommand(`git -C ${initPath} submodule add ${routerModuleUrl} ${routerPath}`);
+		executeCommand(`git -C ${initPath} submodule add ${alphaComponentModuleUrl} ${alphaComponentPath}`);
 	} catch (error) {
 		console.log("error :>> ", error);
 	}
